Upgraded expiry attribute to metadata attribute
The expiry attribute has now become a metadata attribute. It holds several
fields. See Attributes.java. I prepended a version nummer to:

 1. Make this implementation future proof
 2. Ensure that leading zero's are not destroyed

old version are recognized based on the 3 byte length of that field. Using a
cardversion for this is not appropriate since this depends on the issuers and
verifiers.